REFACTOR - PHASE 3

+++ src/pci/pci.h
@@ -29,7 +29,7 @@ typedef struct
 } pci_device_t;
 
 uint32_t pciGetDeviceCount(void);
-int pciGetDevices(pci_device_t *destination, int start, int end);
+size_t pciGetDevices(pci_device_t *destination, size_t start, size_t end);
 const char *pciClassToString(uint8_t class_id, uint8_t subclass_id);
 void pciEnumerateDevices(void (*on_device_found)(pci_device_t *));
 void pciDisplayDeviceInfo(pci_device_t *dev);
